requirements.txt
opencv-python-headless 
deepface
llama-cpp-python 
fastapi
uvicorn[standard]
opencv-python-headless
toml
torch 
torchvision 
pillow 
transformers 
accelerate 
safetensors
bitsandbytes
jinja2
numpy<2
tf-keras
faster-whisper
piper-tts
psutil
langchain
faiss-cpu 
sentence-transformers
pymupdf 
unstructured
langchain-community
langchain-text-splitters
python-multipart
langdetect
sqlalchemy
backup